Adds support for passing password via MPD_HOST
Following what appears to be a standard convention for MPD tools, the
password for an MPD instance can now be supplied via the command line
using the MPD_HOST parameter. If an '@' sign is present in the MPD_HOST
environment variable, then anything before the '@' will be considered
the password.