Add better password support
Before this change, passwords were handled something like this:
1. Run an MPD `status` command.
2. If it fails and the user gave us a password, send the password.
   Otherwise, prompt the user for a password.

If ashuffle was able to run a `status` command successfully, the
password supplied by the user was discarded. As seen in issues like #15,
configurations that gave the default user the `read` permission, but no
other permissions would not use the user's password as expected.

This change introduces a new approach to passwords:
* If a password is supplied by the user, it is always applied.
* The MPD `notcommands` command is used to get a list of commands
  ashuffle is not allowed to run. Instead of just running a `status` to
  determine if it has the appropriate privileges, ashuffle will check
  that no commands it needs to run are present in this list.
* ashuffle will not only prompt for a password if the user did not
  supply one, and ashuffle doesn't have access to a required command.