Adds an extended "test delegate" to shuffle_loop
When testing the core shuffle loop, different tests have different
requirements for reproducibility. For example, tests of the reaction
to MPD events, may want to skip the "initialization" phase of the core
loop. On the other hand, tests of the initialization phase may want to
not run the core loop at all.

The test delegates give tests the ability to do all of these
operations. During normal operation, it the parameter can be set to "NULL"
to use the common case.