Adds support for old MPD releases to integration suite
Before 0.21, MPD was built with a standard `./configure && make`
pattern. This change adds support for these older releases. I've tried
it all the way back to `0.19.12` which is what ashuffle is targeting for
compatibility.

Note: This change also includes a small MPD patch for MPD versions
pre-0.19.15. This is needed to get these old versions of MPD to compile
on our integration test image.