Integration tests discover latest MPD/mpdclient version
Previously, when the MPD or libmpdclient version was set to "latest",
the build scripts would just use a hard-coded "latest" value for the
dependency. With this change, the build scripts will actually pull down
the MPD/libmpdclient repos and parse tag versions to try and discover
the latest version automatically.