Check URIs given with `-f' against local URI db
If `--no-check' is not provided, ashuffle validates that any URIs passed
with `-f' are actually in the MPD library. Before this change, this was
done by querying MPD for the exact user-provided URI. Since we would do
this for every URI given, if a user attempted to shuffle a large number
of songs (10k+) with `-f', ashuffle would take a long time to start.

With this change, instead of querying for each URI individually, we use
the `listall' command (like how build_songs_mpd works) fetch a list of
all valid URIs, then we validate user-provided URIs against this list.
Then, only if a song is in our local list, we check the rules against
MPD.

This should make ashuffle startup much faster for users that:
  A) Are shuffling large sets of songs through using `-f', and
  B) Are not using rules. Rule processing will still be slow, since
     it is done synchronously at startup.