Change ashuffle lint target: s/clang-tidy/ashuffle-clang-target/
Meson 0.52.0 introduce an automatically generated `clang-tidy` target if
the meson root (repo root) contains a clang-tidy config. However, that
target:
  a) Still doesn't work with the meson-generated command lines (-pipe)
  b) Includes a bunch of test files, even though they are not included
     in any rules (since tests are excluded by default).
According to the 0.52.0 patch-notes, meson should not generate the
`clang-tidy` target if a `clang-tidy` target already exists in the meson
config, but *that* doesn't seem to work either. If we leave the
`clang-tidy` target, meson throws an error for two rules with the same
name.

Instead, we rename the existing `clang-tidy` target to
`ashuffle-clang-tidy`. This also updates the linter script to invoke the
renamed target.