Make shuffle chain group-aware.
This patch tweaks the shuffle chain API slightly, to allow adding groups
of URIs in addition to individual URIs. URIs will be shuffled as a
group, and Pick returns all items of the group that was selected. This
should allow us to shuffle by different by groups of songs, instead of
just songs themselves.