Add group_by support to MPD loaders
This change introduces a new "group_by" option on the MPD-based loaders.
When provided, the loaders use these fields to group the songs they read
before adding the items to the ShuffleChain (using the new Shuffle-chain
group support). The grouping is done via a basic unordered_map: Keys are
vectors of tag values corresponding to the group_by fields, and values
are the URIs in the group. Once all Groups are generated, they are added
to the chain.