Wait for MPD DB to update before running tests
Previously, we would not wait for MPD to finish it's initial library
update before running ashuffle against it. This would mean that MPD
would not have a full library, causing tricky problems for our tests.
This change makes the MPD wrapper wait until MPD has finished updating.

However, this also means that MPD now has a full library when the tests
are running, so there were a couple of performance regressions (from
400ms -> 750ms in the 95%-ile) which is unfortunate. At least now we
know we're testing against the full library.