replace unsigned and int with size_t in appropriate places
_pool and _window are used to store indicies of _items,
they should store size_t's just to be sure.
unsigned or int may overflow with really, really big databases
or if the system has a really small (unsigned) int.

It probably won't ever overflow in the wild
(you'd need a few billion songs on most systems),
but theoretically it could and should be made safe.