Prevent absl from being installed
Abseil uses some cmake trickery to figure out if it's being built as
part of another project, or built alone. When built alone, absl
configures itself to be installed into some target library path where it
can be used like a normal library. Meson's cmake integration builds absl
standalone, so absl enters this system mode which Meson picks up on,
marking all absl libraries as `install: true`.

This change avoids this detection in absl by adjusting the
PROJECT_SOURCE_DIR so absl thinks it is part of another project (which
it is) and does not generate installation rules.

Really, it would be nice if we could just pass a `-D` to absl to prevent
it from entering the system mode. I will probably open a PR to do this.

Should fix #57.