integration: Use fakelib instead of `music.huge`.
Fakelib is a library that can generate huge music libraries based off a
single MP3. See joshkunz/fakelib for more. This change uses fakelib
instead of the ashuffle-integration-root built-in `music.huge` library.
This should allow us to scale the fake music library more easily, and
cut down on the build time and size for the integration root image.