meta: Re-tooled Cross-release support.
Meta is currently using the official ARM-released crosstool to do ARM32
ashuffle release builds. This has a couple issues. Namely, the ARM
crosstool only supports armv7 or later, but the RPi Zero is only armv6.
It may also be (or later get-out of) sync with the envrionments users
are using (primarily RPi OS presumably).

This change switches to a Clang/LLVM-based cross build using OS
distribution images themselves as sysroots. This guarantees we will
maintain compatibility with the platforms most of our users are using.
One slight snag of this approach, is that the target distribution images
typically bundle versions of the C++ stdlib that are too old for
ashuffle to use (since we use C++17). To get around this, we actually
build LLVM's libc++ (using the same OS image), and then statically link
it into ashuffle itself.

As of this change, the cross build sysroots are based on:

  armv6h/armv7h: Raspbian 2019-04-02, a fairly old Raspbian.
  arm64: Ubuntu xenial, the oldest supported LTS release.