github: Add config for testing via GitHub actions.
Travis-CI is currently taking ~3hr to run our integration test suite.
This is pretty painful. A lot of this is due to Travis's parallelism
constraints (only 4 jobs at once). In my experience, GitHub actions
have been faster, so lets see if this speeds things up.

This is the first step to switching to GitHub actions, porting the
integration test suite to GitHub actions. Releasing will be trickier,
and will be handled in followup changes.