Update MPD support matrix 2021-11-21
Removes the `Xenial` LTS release versions, as Xenial has reached the end
of the standard suppport window.

Replaces `Groovy` with `Impish` the latest non-LTS release. Technically
`Hirsute` is also still active, but it reaches end of life in January,
and I don't want to slow down the build too much.