test: Build ashuffle before integration tests run
Before this change, the integration test suite would build ashuffle
when the test was started. This wasn't really harmful, but it was yet
another bit of custom code for building ashuffle that would need to be
updated when ashuffle's build steps were changed.

This change removes adds a new meta `testbuild` command that builds a
copy of ashuffle suitable for use in integration tests. This command is
used to build ashuffle as part of the integration container build. The
code to build ashuffle at the beginning of the test is removed.